+++ webpack.config.js
@@ -17,6 +17,10 @@ module.exports = {
 			'./src/blocks/link-page-analytics/index.js',
 		'blocks/link-page-analytics/view':
 			'./src/blocks/link-page-analytics/view.js',
+		'blocks/artist-profile-manager/index':
+			'./src/blocks/artist-profile-manager/index.js',
+		'blocks/artist-profile-manager/view':
+			'./src/blocks/artist-profile-manager/view.js',
 	},
 	output: {
 		...defaultConfig.output,